+++ common/src/main/java/com/ably/tracking/common/workerqueue/Properties.kt
@@ -1,0 +1,8 @@
+package com.ably.tracking.common.workerqueue
+
+/**
+ * An interface containing properties that affect [WorkerQueue] behavior.
+ */
+interface Properties {
+    val isStopped: Boolean
+}