+++ integration-testing-app/src/androidTest/java/com/ably/tracking/tests/PublisherAndSubscriberTests.kt
@@ -4,11 +4,13 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import com.ably.tracking.LocationUpdate
 import com.ably.tracking.TrackableState
+import com.ably.tracking.annotations.Experimental
 import com.ably.tracking.publisher.Trackable
 import com.ably.tracking.subscriber.Subscriber
 import com.ably.tracking.test.android.common.BooleanExpectation
 import com.ably.tracking.test.android.common.UnitExpectation
 import com.ably.tracking.test.android.common.testLogD
+import com.google.common.truth.Truth.assertThat
 import io.ably.lib.realtime.AblyRealtime
 import java.util.UUID
 import kotlinx.coroutines.CoroutineScope
@@ -231,6 +233,9 @@ class PublisherAndSubscriberTests {
         // await
         subscriberFailedExpectation.await()
 
+        // captured before cleanup because currently the subscriber on stop is transitioning the trackable to Offline state, reported here https://github.com/ably/ably-asset-tracking-android/issues/802
+        val finalTrackableState = subscriber.trackableStates.value
+
         // cleanup
         runBlocking {
             subscriber.stop()
@@ -238,5 +243,54 @@ class PublisherAndSubscriberTests {
 
         // then
         subscriberFailedExpectation.assertFulfilled()
+        assertThat(finalTrackableState)
+            .isInstanceOf(TrackableState.Failed::class.java)
+    }
+
+    @OptIn(Experimental::class)
+    @Test
+    fun shouldNotEmitPublisherPresenceFalseIfPublisherIsPresentFromTheStart() {
+        // given
+        val subscriberEmittedPublisherPresentExpectation = UnitExpectation("subscriber emitted publisher present")
+        val context = InstrumentationRegistry.getInstrumentation().targetContext
+        val trackableId = UUID.randomUUID().toString()
+        val scope = CoroutineScope(Dispatchers.Default)
+        val publisherPresentValues = mutableListOf<Boolean>()
+
+        // when
+        // create publisher and start tracking
+        val publisher = createAndStartPublisher(context, sendResolution = true)
+        runBlocking {
+            publisher.track(Trackable(trackableId))
+        }
+
+        // create subscriber and listen for publisher presence
+        var subscriber: Subscriber
+        runBlocking {
+            subscriber = createAndStartSubscriber(trackableId)
+        }
+        subscriber.publisherPresence
+            .onEach { isPublisherPresent ->
+                publisherPresentValues.add(isPublisherPresent)
+                if (isPublisherPresent) {
+                    subscriberEmittedPublisherPresentExpectation.fulfill()
+                }
+            }
+            .launchIn(scope)
+
+        // await for publisher present event
+        subscriberEmittedPublisherPresentExpectation.await()
+
+        // cleanup
+        runBlocking {
+            coroutineScope {
+                launch { publisher.stop() }
+                launch { subscriber.stop() }
+            }
+        }
+
+        // then
+        subscriberEmittedPublisherPresentExpectation.assertFulfilled()
+        Assert.assertTrue("first publisherPresence value should be true", publisherPresentValues.first())
     }
 }