+++ publishing-sdk/src/main/java/com/ably/tracking/publisher/workerqueue/resulthandlers/AddTrackableResultHandler.kt
@@ -18,7 +18,7 @@ internal class AddTrackableResultHandler(
             is AddTrackableWorkResult.Fail ->
                 return workerFactory.createWorker(
                     WorkerParams.AddTrackableFailed(
-                        workResult.trackable, workResult.callbackFunction, workResult.exception as Exception
+                        workResult.trackable, workResult.callbackFunction, workResult.exception as Exception, workResult.isConnectedToAbly
                     )
                 )
 
@@ -29,6 +29,16 @@ internal class AddTrackableResultHandler(
                         workResult.callbackFunction,
                         workResult.presenceUpdateListener,
                         workResult.channelStateChangeListener
+                    )
+                )
+
+            is AddTrackableWorkResult.WorkDelayed ->
+                return workerFactory.createWorker(
+                    WorkerParams.AddTrackable(
+                        workResult.trackable,
+                        workResult.callbackFunction,
+                        workResult.presenceUpdateListener,
+                        workResult.channelStateChangeListener,
                     )
                 )
         }