+++ subscribing-sdk/src/main/java/com/ably/tracking/subscriber/workerqueue/workers/StartConnectionWorker.kt
@@ -2,9 +2,9 @@ package com.ably.tracking.subscriber.workerqueue.workers
 
 import com.ably.tracking.common.Ably
 import com.ably.tracking.common.ResultCallbackFunction
-import com.ably.tracking.subscriber.Properties
+import com.ably.tracking.subscriber.SubscriberProperties
 import com.ably.tracking.subscriber.SubscriberInteractor
-import com.ably.tracking.subscriber.workerqueue.CallbackWorker
+import com.ably.tracking.common.workerqueue.CallbackWorker
 import com.ably.tracking.subscriber.workerqueue.WorkerSpecification
 
 internal class StartConnectionWorker(
@@ -12,25 +12,30 @@ internal class StartConnectionWorker(
     private val subscriberInteractor: SubscriberInteractor,
     private val trackableId: String,
     callbackFunction: ResultCallbackFunction<Unit>
-) : CallbackWorker(callbackFunction) {
+) : CallbackWorker<SubscriberProperties, WorkerSpecification>(callbackFunction) {
     override fun doWork(
-        properties: Properties,
+        properties: SubscriberProperties,
         doAsyncWork: (suspend () -> Unit) -> Unit,
         postWork: (WorkerSpecification) -> Unit
-    ): Properties {
+    ): SubscriberProperties {
         subscriberInteractor.updateTrackableState(properties)
         doAsyncWork {
-            val result = ably.connect(
-                trackableId,
-                properties.presenceData,
-                useRewind = true,
-                willSubscribe = true
-            )
-
-            if (result.isSuccess) {
-                postWork(WorkerSpecification.SubscribeForPresenceMessages(callbackFunction))
+            val startAblyConnectionResult = ably.startConnection()
+            if (startAblyConnectionResult.isFailure) {
+                callbackFunction(startAblyConnectionResult)
             } else {
-                callbackFunction(result)
+                val result = ably.connect(
+                    trackableId,
+                    properties.presenceData,
+                    useRewind = true,
+                    willSubscribe = true
+                )
+
+                if (result.isSuccess) {
+                    postWork(WorkerSpecification.SubscribeForPresenceMessages(callbackFunction))
+                } else {
+                    callbackFunction(result)
+                }
             }
         }
         return properties