+++ test-common/src/main/java/com/ably/tracking/test/common/AblyTestExtensions.kt
@@ -16,6 +16,14 @@ fun Ably.mockCreateConnectionSuccess(trackableId: String) {
     mockSubscribeToPresenceSuccess(trackableId)
 }
 
+fun Ably.mockStartConnectionSuccess() {
+    coEvery { startConnection() } returns Result.success(Unit)
+}
+
+fun Ably.mockStartConnectionFailure() {
+    coEvery { startConnection() } returns Result.failure(anyConnectionException())
+}
+
 fun Ably.mockConnectSuccess(trackableId: String) {
     val callbackSlot = slot<(Result<Unit>) -> Unit>()
     every {
@@ -64,7 +72,7 @@ fun Ably.mockSubscribeToPresenceSuccess(
     } answers {
         callbackSlot.captured(Result.success(Unit))
     }
-    coEvery { subscribeForPresenceMessages(trackableId, capture(listenerSlot)) } returns Result.success(Unit)
+    coEvery { subscribeForPresenceMessages(trackableId, capture(listenerSlot), any<Boolean>()) } returns Result.success(Unit)
 }
 
 fun Ably.mockSubscribeToPresenceError(trackableId: String) {
@@ -74,7 +82,18 @@ fun Ably.mockSubscribeToPresenceError(trackableId: String) {
     } answers {
         callbackSlot.captured(Result.failure(anyConnectionException()))
     }
-    coEvery { subscribeForPresenceMessages(trackableId, any()) } returns Result.failure(anyConnectionException())
+    coEvery { subscribeForPresenceMessages(trackableId, any(), any<Boolean>()) } returns Result.failure(anyConnectionException())
+}
+
+fun Ably.mockGetCurrentPresenceSuccess(
+    trackableId: String,
+    currentPresenceMessage: List<PresenceMessage> = emptyList(),
+) {
+    coEvery { getCurrentPresence(trackableId) } returns Result.success(currentPresenceMessage)
+}
+
+fun Ably.mockGetCurrentPresenceError(trackableId: String) {
+    coEvery { getCurrentPresence(trackableId) } returns Result.failure(anyConnectionException())
 }
 
 fun Ably.mockDisconnect(trackableId: String, result: Result<Unit>) {